+++ capi/example/soem.c
@@ -4,7 +4,7 @@
  * Created Date: 16/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 16/05/2022
+ * Last Modified: 24/05/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Hapis Lab. All rights reserved.
@@ -56,6 +56,8 @@ int main() {
     free(error);
     return -1;
   }
+
+  AUTDSetCheckAck(cnt, true);
 
   return run(cnt);
 }