+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.2
+* Add `DynamicTransducer` for C-API
+* Remove legacy C-API library
+* Change `AmplitudeConstraint` api in Holo Gain in C-API
+* Fix `wavelength` and `wavenumber` of `NormalTransducer`
+
 # 2.0.1
 * Fix C-API
 * Add `objective` parameter to `gain::holo::Greedy`